Extract table data query logic
