+++ examples/recipes/src/data_viewer/DataViewer.tsx
@@ -2,11 +2,12 @@ import { Box, Paper } from '@mui/material'
 import { useLiveQuery } from 'electric-sql/react'
 import { useElectric } from '../electric/ElectricWrapper'
 import { useMemo, useState } from 'react'
-import { ColumnDef, PaginationState, SortingState, TableView } from './TableView'
+import { ColumnDef, TableView } from './TableView'
 import { columns, aggregateColumns } from './commerce_orders_columns'
 import { QueryBuilder } from './QueryBuilder'
 import { ChartView } from './ChartView'
 import { Selector } from './Selector'
+import { PaginationState, SortingState, useTableData } from './use_table_data'
 
 interface PropertyValue {
   property: string
@@ -51,45 +52,14 @@ const TableDataViewer = ({
 
   // Keep an order by clause generated by the sorting of columns
   const [sorting, setSorting] = useState<SortingState[]>([])
-  const orderByClause = useMemo(() => {
-    const sortStatements = sorting
-      .filter((sortState) => !!sortState.order)
-      .map((sortState) => `${sortState.field} ${sortState.order}`)
-    return sortStatements.length > 0 ? `ORDER BY ${sortStatements.join(',')}` : ''
-  }, [sorting])
 
-  const { db } = useElectric()!
-
-  // Get the order data for the given query
-  const { results: orders = [] } = useLiveQuery(
-    db.liveRawQuery({
-      sql: `
-      SELECT * FROM commerce_orders
-      WHERE ${whereClause}
-      ${orderByClause}
-      LIMIT ${pagination.pageSize}
-      OFFSET ${pagination.pageIndex * pagination.pageSize}
-    `,
-    }),
-  )
-
-  // Also get a count for the total data matching the filters
-  // to inform the pagination controller
-  const totalNumOrders =
-    useLiveQuery(
-      db.liveRawQuery({
-        sql: `
-      SELECT COUNT(*) FROM commerce_orders
-      WHERE ${whereClause};
-    `,
-      }),
-    ).results?.[0]?.['COUNT(*)'] ?? 0
+  const { orders, totalNumberOfOrders } = useTableData({ sorting, pagination, whereClause })
 
   return (
     <TableView
       columns={columns}
       rows={orders}
-      totalNumberOfRows={totalNumOrders}
+      totalNumberOfRows={totalNumberOfOrders}
       pagination={pagination}
       onPaginationChange={setPagination}
       sorting={sorting}
